make gobpgd buildable on Mac OS X
Not tested on Mac OS X but at least buildable.
